+++ CHANGES.rst
@@ -4,12 +4,13 @@ CHANGES
 `Unreleased <https://github.com/fmigneault/aiu/tree/master>`_ (latest)
 ------------------------------------------------------------------------------------
 
-* Nothing yet.
+* Add support of YouTube Music playlist links corresponding to singles and EPs "pseudo-albums" of an artist.
+* Modernize Python project and linting tools.
 
 `2.0.4 <https://github.com/fmigneault/aiu/tree/2.0.4>`_ (2025-03-22)
 ------------------------------------------------------------------------------------
 
-* Add more gracefull error handling and warning message if no albums are found for the provided YouTube Music link.
+* Add more graceful error handling and warning message if no albums are found for the provided YouTube Music link.
 
 `2.0.3 <https://github.com/fmigneault/aiu/tree/2.0.3>`_ (2025-03-22)
 ------------------------------------------------------------------------------------
@@ -18,7 +19,7 @@ CHANGES
 * Replace ``eyeD3==0.9.6``
   by `fmigneault/eyeD3@no-install-tests <https://github.com/fmigneault/eyeD3/tree/no-install-tests>`_
   to avoid invalid extra tests dependencies.
-* Fix `CR/LF` across repository using portable `LF` everywhere.
+* Fix ``CR/LF`` across repository using portable ``LF`` everywhere.
 * Fix multiple linting issues with additional linting tools.
 
 `2.0.2 <https://github.com/fmigneault/aiu/tree/2.0.2>`_ (2025-03-17)